Network: Split base_classes into smaller modules
