+++ src/net/builders.py
@@ -1,4 +1,3 @@
-import abc
 import re
 
 import tflearn
@@ -9,151 +8,6 @@ import tflearn.layers.merge_ops as merge
 import tflearn.layers.estimator as est
 
 import net.constants as net_cons
-import net.layer.layer_utils as layer_utils
-
-
-class NetworkModel:
-
-    def __init__(self, neural_network, **model_settings):
-        self._net = neural_network
-        self._hidden_models = {}
-        self.initialize_model(**model_settings)
-
-    def _update_hidden_models(self):
-        hidden_models = self._hidden_models
-        session = self._model.session
-        for layer_name, model in hidden_models.items():
-            model.session.close()
-            model.session = session
-            model.trainer.session = session
-            model.predictor.session = session
-
-    def get_layer_weights(self, layer_name):
-        layer = self._net.trainable_layers[layer_name]
-        converter = layer_utils.weight_converters_external().get(
-            layer['type'], lambda w: w)
-        return converter(self._model.get_weights(layer['layer'].W))
-
-    def set_layer_weights(self, layer_name, weights):
-        layer = self._net.trainable_layers[layer_name]
-        converter = layer_utils.weight_converters_internal().get(
-            layer['type'], lambda w: w)
-        self._model.set_weights(layer['layer'].W, converter(weights))
-
-    def get_layer_biases(self, layer_name):
-        layer = self._net.trainable_layers[layer_name]
-        return self._model.get_weights(layer['layer'].b)
-
-    def set_layer_biases(self, layer_name, weights):
-        layer = self._net.trainable_layers[layer_name]
-        self._model.set_weights(layer['layer'].b, weights)
-
-    @property
-    def network_graph(self):
-        return self._net
-
-    @property
-    def network_model(self):
-        return self._model
-
-    def initialize_model(self, create_hidden_models=False, **model_settings):
-        settings = {
-            'tensorboard_dir': model_settings.get('tb_dir',
-                                                  '/tmp/tflearn_logs/'),
-            'tensorboard_verbose': model_settings.get('tb_verbosity', 0)
-        }
-        model = tflearn.DNN(self._net.output_layer['layer'], **settings)
-        self._model = model
-        if create_hidden_models:
-            session = self._model.session
-            layers = set()
-            for pathname, path in self._net.data_paths.items():
-                layers = layers.union(path)
-            layers.remove(self._net.output_layer)
-            hidden_layers = self._net.hidden_layers
-            hidden_models = {name: tflearn.DNN(hidden_layers[name]['layer'],
-                                               session=session)
-                             for name in layers}
-            self._hidden_models = hidden_models
-
-    def get_hidden_layer_activations(self, input_data_dict):
-        if len(self._hidden_models) == 0:
-            raise Exception('Hidden layer activations not retrievable.')
-        hidden_models = self._hidden_models
-        return {layer_name: model.predict(input_data_dict)
-                for layer_name, model in hidden_models.items()}
-
-    def load_from_file(self, model_filename, **optargs):
-        w_only = optargs.get('weights_only', False)
-        self._model.load(model_filename, weights_only=w_only)
-        if len(self._hidden_models) > 0:
-            self._update_hidden_models()
-
-    def save_to_file(self, model_filename):
-        self._model.save(model_filename)
-
-
-class NeuralNetwork(abc.ABC):
-
-    # legal names contain alphanumeric characters, underscore, dash and dots
-    # (essentially, filesystem friendly)
-    __allowable_layer_name_regex__ = re.compile('[a-zA-Z0-9_\\-.]+')
-    __default_layer_name_regex__ = re.compile('([^/]+)/.+')
-
-    def __init__(self, builder):
-        layers, categories = builder.layers_dict, builder.layer_categories
-        trainable_layers = categories['trainable']
-        hidden_layers = categories['hidden']
-        input_layers = categories['input']
-        self._out = layers[builder.output_layer]
-        self._out_name = builder.output_layer
-        self._inputs = {name: layers[name] for name in input_layers}
-        self._trainable = {
-            name: layers[name] for name in trainable_layers
-        }
-        self._hidden = {name: layers[name] for name in hidden_layers}
-        self._paths = builder.data_paths.copy()
-
-    # properties
-
-    @property
-    @abc.abstractmethod
-    def network_type(self):
-        pass
-
-    @property
-    @abc.abstractmethod
-    def input_spec(self):
-        pass
-
-    @property
-    @abc.abstractmethod
-    def output_spec(self):
-        pass
-
-    @property
-    def trainable_layers(self):
-        return self._trainable
-
-    @property
-    def hidden_layers(self):
-        return self._hidden
-
-    @property
-    def input_layers(self):
-        return self._inputs
-
-    @property
-    def output_layer(self):
-        return self._out
-
-    @property
-    def output_layer_name(self):
-        return self._out_name
-
-    @property
-    def data_paths(self):
-        return self._paths
 
 
 class GraphBuilder:
