+++ src/net/samples/github_net3.py
@@ -2,21 +2,23 @@
 # github_net with added local_response_normalization layers after max pooling
 # and removed dropout and flatten layers before first fc layer
 
-import net.base_classes as base_classes
+import net.builders as builders
+import net.graphs as graphs
+import net.models as models
 
 
 def create_model(input_shapes, **optsettings):
     network = GithubNet3(input_shapes, **optsettings)
-    return base_classes.NetworkModel(network, **optsettings)
+    return models.NetworkModel(network, **optsettings)
 
 
-class GithubNet3(base_classes.NeuralNetwork):
+class GithubNet3(graphs.NeuralNetwork):
 
     def __init__(self, input_shapes, input_type='yx', **optsettings):
         lr = optsettings.get('learning_rate') or 0.001
         optimizer = optsettings.get('optimizer') or 'adam'
         loss_fn = optsettings.get('loss_fn') or 'categorical_crossentropy'
-        builder = base_classes.GraphBuilder()
+        builder = builders.GraphBuilder()
         shape = input_shapes[input_type]
         builder.add_input_layer(shape, input_type, name='input')
         builder.start_new_path()
