+++ src/net/samples/mnist.py
@@ -7,21 +7,23 @@ Links:
     [MNIST Dataset] http://yann.lecun.com/exdb/mnist/
 """
 
-import net.base_classes as base_classes
+import net.builders as builders
+import net.graphs as graphs
+import net.models as models
 
 
 def create_model(input_shapes, **optsettings):
     network = MNISTNet(input_shapes, **optsettings)
-    return base_classes.NetworkModel(network, **optsettings)
+    return models.NetworkModel(network, **optsettings)
 
 
-class MNISTNet(base_classes.NeuralNetwork):
+class MNISTNet(graphs.NeuralNetwork):
 
     def __init__(self, input_shapes, input_type='yx', **optsettings):
         lr = optsettings.get('learning_rate') or 0.01
         optimizer = optsettings.get('optimizer') or 'adam'
         loss_fn = optsettings.get('loss_fn') or 'categorical_crossentropy'
-        builder = base_classes.GraphBuilder()
+        builder = builders.GraphBuilder()
         shape = input_shapes[input_type]
         builder.add_input_layer(shape, input_type, name='input')
         builder.start_new_path()
